Reverted param to 'db_iterator' from 'tensor_db' for AggregationFunction interface. Moved PrivilegedAggregationFunction to experimental directory
