+++ openfl/component/aggregation_functions/__init__.py
@@ -5,7 +5,8 @@
 
 from .adagrad_adaptive_aggregation import AdagradAdaptiveAggregation
 from .adam_adaptive_aggregation import AdamAdaptiveAggregation
-from .core import AggregationFunction, PrivilegedAggregationFunction
+from .core import AggregationFunction
+from .experimental import PrivilegedAggregationFunction
 from .fedcurv_weighted_average import FedCurvWeightedAverage
 from .geometric_median import GeometricMedian
 from .median import Median
