+++ openfl/component/aggregation_functions/experimental/__init__.py
@@ -1,0 +1,8 @@
+# Copyright (C) 2020-2021 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+"""Aggregation functions experimental package."""
+
+from .privileged_aggregation import PrivilegedAggregationFunction
+
+__all__ = ['PrivilegedAggregationFunction',]
